+++ test/main.test.js
@@ -244,6 +244,25 @@ describe('Set-DOM', function () {
     assert.equal(el1.innerHTML, '<div class="a" data-checksum="abc">initial</div>', 'did nothing')
   })
 
+  it('should diff children (data-checksum) custom attribute', function () {
+    // Set custom checksum attribute
+    diff.CHECKSUM = 'data-custom-checksum'
+
+    var el1 = document.createElement('div')
+    var el2 = document.createElement('div')
+
+    // Update inner html
+    el1.innerHTML = '<div class="a" data-custom-checksum="abc">initial</div>'
+    el2.innerHTML = '<div class="b" data-custom-checksum="efg">final</div>'
+
+    // Attempt to diff
+    diff(el1, el2)
+    assert.equal(el1.innerHTML, '<div class="b" data-custom-checksum="efg">final</div>', 'should have updated element')
+
+    // Reset custom checksum attribute
+    diff.CHECKSUM = 'data-checksum'
+  })
+
   it('should automatically parse html for diff', function () {
     var el = document.createElement('div')